Rename multi_distinct_to_union.rs to multi_distinct_to_cross_join.rs
