+++ datafusion/optimizer/src/multi_distinct_to_cross_join.rs
