+++ datafusion/optimizer/src/optimizer.rs
@@ -45,7 +45,7 @@ use crate::eliminate_outer_join::EliminateOuterJoin;
 use crate::extract_equijoin_predicate::ExtractEquijoinPredicate;
 use crate::extract_leaf_expressions::{ExtractLeafExpressions, PushDownLeafProjections};
 use crate::filter_null_join_keys::FilterNullJoinKeys;
-use crate::multi_distinct_to_union::MultiDistinctToCrossJoin;
+use crate::multi_distinct_to_cross_join::MultiDistinctToCrossJoin;
 use crate::optimize_projections::OptimizeProjections;
 use crate::optimize_unions::OptimizeUnions;
 use crate::plan_signature::LogicalPlanSignature;
